Move method
The getCommandObjectBindingSourceForPrefix method is a utility method required from just one place (a Controller).
